Refactor to bedrock package
