+++ src/main/java/com/viaversion/viafabricplus/injection/mixin/features/bedrock/block/shape/MixinConduitBlock.java
@@ -19,14 +19,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.viaversion.viafabricplus.injection.mixin.features.block.shape;
+package com.viaversion.viafabricplus.injection.mixin.features.bedrock.block.shape;
 
 import com.viaversion.viafabricplus.protocoltranslator.ProtocolTranslator;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.SeaPickleBlock;
+import net.minecraft.block.ConduitBlock;
 import net.minecraft.block.ShapeContext;
-import net.minecraft.state.property.IntProperty;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.util.shape.VoxelShapes;
@@ -40,37 +39,21 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(SeaPickleBlock.class)
-public class MixinSeaPickleBlock extends Block {
+@Mixin(ConduitBlock.class)
+public abstract class MixinConduitBlock extends Block {
 
     @Unique
-    private static final VoxelShape viaFabricPlus$shape_bedrock = Block.createColumnShape(16.0F, 0.0F, 6.0F);
+    private static final VoxelShape viaFabricPlus$shape_bedrock = VoxelShapes.cuboid(0.25, 0, 0.25, 0.75, 0.5, 0.75);
 
     @Shadow
     @Final
-    private static  VoxelShape ONE_PICKLE_SHAPE;
+    private static VoxelShape SHAPE;
 
-    @Shadow
-    @Final
-    private static  VoxelShape TWO_PICKLES_SHAPE;
-
-    @Shadow
-    @Final
-    private static  VoxelShape THREE_PICKLES_SHAPE;
-
-    @Shadow
-    @Final
-    private static  VoxelShape FOUR_PICKLES_SHAPE;
-
-    @Shadow
-    @Final
-    public static IntProperty PICKLES;
-
-    public MixinSeaPickleBlock(final Settings settings) {
+    public MixinConduitBlock(final Settings settings) {
         super(settings);
     }
 
-    @Inject(method = "getOutlineShape", at = @At("HEAD"), cancellable = true)
+    @Inject(method = "getOutlineShape", at = @At(value = "RETURN"), cancellable = true)
     private void changeOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context, CallbackInfoReturnable<VoxelShape> cir) {
         if (ProtocolTranslator.getTargetVersion().equals(BedrockProtocolVersion.bedrockLatest)) {
             cir.setReturnValue(viaFabricPlus$shape_bedrock);
@@ -78,25 +61,12 @@ public class MixinSeaPickleBlock extends Block {
     }
 
     @Override
-    public VoxelShape getCollisionShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        if (ProtocolTranslator.getTargetVersion().equals(BedrockProtocolVersion.bedrockLatest)) {
-            return VoxelShapes.empty();
-        } else {
-            return super.getCollisionShape(state, world, pos, context);
-        }
-    }
-
-    @Override
     public VoxelShape getCullingShape(BlockState state) {
         if (ProtocolTranslator.getTargetVersion().equals(BedrockProtocolVersion.bedrockLatest)) {
-            return switch (state.get(PICKLES)) {
-                case 2 -> TWO_PICKLES_SHAPE;
-                case 3 -> THREE_PICKLES_SHAPE;
-                case 4 -> FOUR_PICKLES_SHAPE;
-                default -> ONE_PICKLE_SHAPE;
-            };
+            return SHAPE;
         } else {
             return super.getCullingShape(state);
         }
     }
+
 }
