+++ src/main/java/com/viaversion/viafabricplus/injection/mixin/features/bedrock/block/shape/MixinDoorBlock.java
@@ -19,51 +19,60 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.viaversion.viafabricplus.injection.mixin.features.block.shape;
+package com.viaversion.viafabricplus.injection.mixin.features.bedrock.block.shape;
 
 import com.viaversion.viafabricplus.protocoltranslator.ProtocolTranslator;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.CakeBlock;
-import net.minecraft.block.CandleCakeBlock;
-import net.minecraft.block.ShapeContext;
-import net.minecraft.util.math.BlockPos;
+import net.minecraft.block.DoorBlock;
+import net.minecraft.state.property.EnumProperty;
+import net.minecraft.util.math.Direction;
 import net.minecraft.util.shape.VoxelShape;
-import net.minecraft.world.BlockView;
+import net.minecraft.util.shape.VoxelShapes;
 import net.raphimc.viabedrock.api.BedrockProtocolVersion;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import org.spongepowered.asm.mixin.injection.Redirect;
+import java.util.Map;
 
-@Mixin(CandleCakeBlock.class)
-public abstract class MixinCandleCakeBlock extends Block {
+@Mixin(DoorBlock.class)
+public abstract class MixinDoorBlock extends Block {
 
     @Unique
-    private static final VoxelShape viaFabricPlus$shape_bedrock = Block.createColumnShape(14.0F, 0.0F, 8.0F);
+    private static final Map<Direction, VoxelShape> viaFabricPlus$shape_bedrock = Map.of(
+        Direction.NORTH, VoxelShapes.cuboid(0, 0, 0.8175, 1, 1, 1),
+        Direction.SOUTH, VoxelShapes.cuboid(0, 0, 0, 1, 1, 0.1825),
+        Direction.WEST, VoxelShapes.cuboid(0.8175, 0, 0, 1, 1, 1),
+        Direction.EAST, VoxelShapes.cuboid(0, 0, 0, 0.1825, 1, 1)
+    );
 
     @Shadow
     @Final
-    private static VoxelShape SHAPE;
+    private static Map<Direction, VoxelShape> SHAPES_BY_DIRECTION;
 
-    public MixinCandleCakeBlock(final Settings settings) {
+    @Shadow
+    @Final
+    public static EnumProperty<Direction> FACING;
+
+    public MixinDoorBlock(final Settings settings) {
         super(settings);
     }
 
-    @Inject(method = "getOutlineShape", at = @At(value = "HEAD"), cancellable = true)
-    private void changeOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context, CallbackInfoReturnable<VoxelShape> cir) {
+    @Redirect(method = "getOutlineShape", at = @At(value = "FIELD", target = "Lnet/minecraft/block/DoorBlock;SHAPES_BY_DIRECTION:Ljava/util/Map;"))
+    private Map<Direction, VoxelShape> changeOutlineShape() {
         if (ProtocolTranslator.getTargetVersion().equals(BedrockProtocolVersion.bedrockLatest)) {
-            cir.setReturnValue(viaFabricPlus$shape_bedrock);
+            return viaFabricPlus$shape_bedrock;
         }
+        return SHAPES_BY_DIRECTION;
     }
 
     @Override
     public VoxelShape getCullingShape(BlockState state) {
         if (ProtocolTranslator.getTargetVersion().equals(BedrockProtocolVersion.bedrockLatest)) {
-            return SHAPE;
+            return SHAPES_BY_DIRECTION.get(state.get(FACING));
         } else {
             return super.getCullingShape(state);
         }
