+++ src/main/java/com/viaversion/viafabricplus/injection/mixin/features/bedrock/block/shape/MixinDragonEggBlock.java
@@ -19,12 +19,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.viaversion.viafabricplus.injection.mixin.features.block.shape;
+package com.viaversion.viafabricplus.injection.mixin.features.bedrock.block.shape;
 
 import com.viaversion.viafabricplus.protocoltranslator.ProtocolTranslator;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.ConduitBlock;
+import net.minecraft.block.DragonEggBlock;
 import net.minecraft.block.ShapeContext;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.shape.VoxelShape;
@@ -34,29 +34,25 @@ import net.raphimc.viabedrock.api.BedrockProtocolVersion;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(ConduitBlock.class)
-public abstract class MixinConduitBlock extends Block {
-
-    @Unique
-    private static final VoxelShape viaFabricPlus$shape_bedrock = VoxelShapes.cuboid(0.25, 0, 0.25, 0.75, 0.5, 0.75);
+@Mixin(DragonEggBlock.class)
+public abstract class MixinDragonEggBlock extends Block {
 
     @Shadow
     @Final
     private static VoxelShape SHAPE;
 
-    public MixinConduitBlock(final Settings settings) {
+    public MixinDragonEggBlock(final Settings settings) {
         super(settings);
     }
 
     @Inject(method = "getOutlineShape", at = @At(value = "RETURN"), cancellable = true)
     private void changeOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context, CallbackInfoReturnable<VoxelShape> cir) {
         if (ProtocolTranslator.getTargetVersion().equals(BedrockProtocolVersion.bedrockLatest)) {
-            cir.setReturnValue(viaFabricPlus$shape_bedrock);
+            cir.setReturnValue(VoxelShapes.fullCube());
         }
     }
 
@@ -68,5 +64,4 @@ public abstract class MixinConduitBlock extends Block {
             return super.getCullingShape(state);
         }
     }
-
 }
