+++ src/main/java/com/viaversion/viafabricplus/injection/mixin/features/bedrock/block/shape/MixinSeaPickleBlock.java
@@ -19,63 +19,84 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.viaversion.viafabricplus.injection.mixin.features.block.shape;
+package com.viaversion.viafabricplus.injection.mixin.features.bedrock.block.shape;
 
 import com.viaversion.viafabricplus.protocoltranslator.ProtocolTranslator;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.DoorBlock;
-import net.minecraft.state.property.EnumProperty;
-import net.minecraft.util.math.Direction;
+import net.minecraft.block.SeaPickleBlock;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.state.property.IntProperty;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.util.shape.VoxelShapes;
+import net.minecraft.world.BlockView;
 import net.raphimc.viabedrock.api.BedrockProtocolVersion;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
-import java.util.Map;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(DoorBlock.class)
-public abstract class MixinDoorBlock extends Block {
+@Mixin(SeaPickleBlock.class)
+public class MixinSeaPickleBlock extends Block {
 
     @Unique
-    private static final Map<Direction, VoxelShape> viaFabricPlus$shape_bedrock = Map.of(
-        Direction.NORTH, VoxelShapes.cuboid(0, 0, 0.8175, 1, 1, 1),
-        Direction.SOUTH, VoxelShapes.cuboid(0, 0, 0, 1, 1, 0.1825),
-        Direction.WEST, VoxelShapes.cuboid(0.8175, 0, 0, 1, 1, 1),
-        Direction.EAST, VoxelShapes.cuboid(0, 0, 0, 0.1825, 1, 1)
-    );
+    private static final VoxelShape viaFabricPlus$shape_bedrock = Block.createColumnShape(16.0F, 0.0F, 6.0F);
 
     @Shadow
     @Final
-    private static Map<Direction, VoxelShape> SHAPES_BY_DIRECTION;
+    private static  VoxelShape ONE_PICKLE_SHAPE;
 
     @Shadow
     @Final
-    public static EnumProperty<Direction> FACING;
+    private static  VoxelShape TWO_PICKLES_SHAPE;
 
-    public MixinDoorBlock(final Settings settings) {
+    @Shadow
+    @Final
+    private static  VoxelShape THREE_PICKLES_SHAPE;
+
+    @Shadow
+    @Final
+    private static  VoxelShape FOUR_PICKLES_SHAPE;
+
+    @Shadow
+    @Final
+    public static IntProperty PICKLES;
+
+    public MixinSeaPickleBlock(final Settings settings) {
         super(settings);
     }
 
-    @Redirect(method = "getOutlineShape", at = @At(value = "FIELD", target = "Lnet/minecraft/block/DoorBlock;SHAPES_BY_DIRECTION:Ljava/util/Map;"))
-    private Map<Direction, VoxelShape> changeOutlineShape() {
+    @Inject(method = "getOutlineShape", at = @At("HEAD"), cancellable = true)
+    private void changeOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context, CallbackInfoReturnable<VoxelShape> cir) {
         if (ProtocolTranslator.getTargetVersion().equals(BedrockProtocolVersion.bedrockLatest)) {
-            return viaFabricPlus$shape_bedrock;
+            cir.setReturnValue(viaFabricPlus$shape_bedrock);
         }
-        return SHAPES_BY_DIRECTION;
+    }
+
+    @Override
+    public VoxelShape getCollisionShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        if (ProtocolTranslator.getTargetVersion().equals(BedrockProtocolVersion.bedrockLatest)) {
+            return VoxelShapes.empty();
+        } else {
+            return super.getCollisionShape(state, world, pos, context);
+        }
     }
 
     @Override
     public VoxelShape getCullingShape(BlockState state) {
         if (ProtocolTranslator.getTargetVersion().equals(BedrockProtocolVersion.bedrockLatest)) {
-            return SHAPES_BY_DIRECTION.get(state.get(FACING));
+            return switch (state.get(PICKLES)) {
+                case 2 -> TWO_PICKLES_SHAPE;
+                case 3 -> THREE_PICKLES_SHAPE;
+                case 4 -> FOUR_PICKLES_SHAPE;
+                default -> ONE_PICKLE_SHAPE;
+            };
         } else {
             return super.getCullingShape(state);
         }
     }
-
 }
