+++ src/main/java/com/viaversion/viafabricplus/injection/mixin/features/movement/constants/MixinEntity.java
