+++ src/main/resources/viafabricplus.mixins.json
@@ -46,6 +46,16 @@
     "features.april_fools_8bit_sound.MixinStaticSound",
     "features.bedrock.allow_new_line.MixinTextRenderer",
     "features.bedrock.reach_around_raycast.MixinGameRenderer",
+    "features.bedrock.block.shape.MixinCactusBlock",
+    "features.bedrock.block.shape.MixinCandleCakeBlock",
+    "features.bedrock.block.shape.MixinConduitBlock",
+    "features.bedrock.block.shape.MixinDoorBlock",
+    "features.bedrock.block.shape.MixinDragonEggBlock",
+    "features.bedrock.block.shape.MixinHoneyBlock",
+    "features.bedrock.block.shape.MixinLanternBlock",
+    "features.bedrock.block.shape.MixinLecternBlock",
+    "features.bedrock.block.shape.MixinSeaPickleBlock",
+    "features.bedrock.block.shape.MixinTrapdoorBlock",
     "features.bedrock.remove_dummy_player_suggestions.MixinClientCommandSource",
     "features.block.interaction.MixinAbstractSignBlock",
     "features.block.interaction.MixinCanPlaceAt1_14",
@@ -61,15 +71,10 @@
     "features.block.shape.MixinAnvilBlock",
     "features.block.shape.MixinBedBlock",
     "features.block.shape.MixinBrewingStandBlock",
-    "features.block.shape.MixinCactusBlock",
-    "features.block.shape.MixinCandleCakeBlock",
     "features.block.shape.MixinCarpetBlock",
     "features.block.shape.MixinCauldronBlock",
     "features.block.shape.MixinChestBlock",
-    "features.block.shape.MixinConduitBlock",
     "features.block.shape.MixinCropBlocks",
-    "features.block.shape.MixinDoorBlock",
-    "features.block.shape.MixinDragonEggBlock",
     "features.block.shape.MixinEnderChestBlock",
     "features.block.shape.MixinEndPortalBlock",
     "features.block.shape.MixinEndPortalFrameBlock",
@@ -78,22 +83,17 @@
     "features.block.shape.MixinFenceGateBlock",
     "features.block.shape.MixinFireBlock",
     "features.block.shape.MixinFlowerbedBlock",
-    "features.block.shape.MixinHoneyBlock",
     "features.block.shape.MixinHopperBlock",
     "features.block.shape.MixinHorizontalConnectingBlock",
     "features.block.shape.MixinLadderBlock",
-    "features.block.shape.MixinLanternBlock",
     "features.block.shape.MixinLeavesBlock",
-    "features.block.shape.MixinLecternBlock",
     "features.block.shape.MixinLilyPadBlock",
     "features.block.shape.MixinPaneBlock",
     "features.block.shape.MixinPistonBlock",
     "features.block.shape.MixinPistonHeadBlock",
-    "features.block.shape.MixinSeaPickleBlock",
     "features.block.shape.MixinPitcherCropBlock",
     "features.block.shape.MixinSnowBlock",
     "features.block.shape.MixinSoulSandBlock",
-    "features.block.shape.MixinTrapdoorBlock",
     "features.block.shape.MixinWallBlock",
     "features.classic.cpe_extension.MixinClassicProtocolExtension",
     "features.classic.cpe_extension.MixinClientboundPacketsc0_30cpe",
